Updated implementations and tests after rework
