+++ test/unit/Exception/ModuleRunExceptionTest.php
@@ -30,6 +30,8 @@ class ModuleRunExceptionTest extends TestCase
 
     /**
      * @since [*next-version*]
+     *
+     * @throws TestSubject
      */
     public function testCanBeThrown()
     {
