+++ test/unit/FactoryInterfaceTest.php
@@ -3,6 +3,7 @@
 namespace Dhii\Modular\UnitTest\Module;
 
 use Dhii\Modular\Module\FactoryInterface as TestSubject;
+use Dhii\Modular\Module\ServiceInterface;
 use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
 
@@ -35,5 +36,6 @@ class FactoryInterfaceTest extends TestCase
         $subject = $this->createInstance();
 
         $this->assertInstanceOf(TestSubject::class, $subject);
+        $this->assertInstanceOf(ServiceInterface::class, $subject);
     }
 }
