+++ test/unit/Services/ExtensionTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Dhii\Modular\Module\UnitTest\Factories;
+namespace Dhii\Modular\Module\UnitTest\Services;
 
-use Dhii\Modular\Module\Factories\Service as TestSubject;
-use Dhii\Modular\Module\FactoryInterface;
+use Dhii\Modular\Module\ExtensionInterface;
+use Dhii\Modular\Module\Services\Extension as TestSubject;
 use Dhii\Modular\Module\Test\CallbackInterface;
 use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
@@ -12,7 +12,7 @@ use Psr\Container\ContainerInterface;
 /**
  * @since [*next-version*]
  */
-class ServiceTest extends TestCase
+class ExtensionTest extends TestCase
 {
     /**
      * @since [*next-version*]
@@ -21,7 +21,7 @@ class ServiceTest extends TestCase
     {
         $subject = new TestSubject([], function () {});
 
-        static::assertInstanceOf(FactoryInterface::class, $subject);
+        static::assertInstanceOf(ExtensionInterface::class, $subject);
     }
 
     /**
@@ -43,7 +43,7 @@ class ServiceTest extends TestCase
     /**
      * @since [*next-version*]
      */
-    public function testInvoke()
+    public function testInvokeNoPrev()
     {
         {
             $dep1 = uniqid('dep1');
@@ -59,7 +59,7 @@ class ServiceTest extends TestCase
             $callback = $this->getMockForAbstractClass(CallbackInterface::class);
             $callback->expects(static::once())
                      ->method('__invoke')
-                     ->with($service1, $service2)
+                     ->with(null, $service1, $service2)
                      ->willReturn($result);
         }
         {
@@ -74,5 +74,43 @@ class ServiceTest extends TestCase
         $subject = new TestSubject([$dep1, $dep2], $callback);
 
         self::assertSame($result, $subject($container));
+    }
+
+    /**
+     * @since [*next-version*]
+     */
+    public function testInvokeWithPrev()
+    {
+        {
+            $prev = uniqid('prev');
+
+            $dep1 = uniqid('dep1');
+            $dep2 = uniqid('dep2');
+
+            $service1 = uniqid('service1');
+            $service2 = uniqid('service2');
+
+            $result = uniqid('result');
+        }
+        {
+            /* @var $callback MockObject|callable */
+            $callback = $this->getMockForAbstractClass(CallbackInterface::class);
+            $callback->expects(static::once())
+                     ->method('__invoke')
+                     ->with($prev, $service1, $service2)
+                     ->willReturn($result);
+        }
+        {
+            /* @var $container MockObject|ContainerInterface */
+            $container = $this->getMockForAbstractClass(ContainerInterface::class);
+            $container->expects(static::exactly(2))
+                      ->method('get')
+                      ->withConsecutive([$dep1], [$dep2])
+                      ->willReturnOnConsecutiveCalls($service1, $service2);
+        }
+
+        $subject = new TestSubject([$dep1, $dep2], $callback);
+
+        self::assertSame($result, $subject($container, $prev));
     }
 }
